Prevent assumed tid translation [Fix #5]
When iterating over an input file to search for tids that may
require translation, if the end of the file is reached with no
matches, `i` is equal to the number of `file_entries`.

It appears an assumption that at least one translation will be
required and `trans[i] = j` is executed even in this case.
As `trans` is only equal to the length of `file_entries - 1`,
updating the contents of `trans` outside of the array is an
invalid operation and appears to cause a memory error that
prevents the `linepointer` buffer used with `getline` to be
free'd safely.

I check here whether `i < file_entries`, *i.e.*, a match was
found, only then is the 'j-loop' executed and `trans` updated.
Note here I make the assumption that the correct `j` will be
found (and so don't check `j < replace_entries`).